Migrate all Jvm Process functions to internal visibility
