+++ library/core-resource/api/core-resource.api
@@ -85,7 +85,7 @@ public final class io/matthewnelson/kmp/tor/core/resource/OSHost$Windows : io/ma
 public final class io/matthewnelson/kmp/tor/core/resource/OSInfo {
 	public static final field Companion Lio/matthewnelson/kmp/tor/core/resource/OSInfo$Companion;
 	public static final field INSTANCE Lio/matthewnelson/kmp/tor/core/resource/OSInfo;
-	public synthetic fun <init> (Lio/matthewnelson/kmp/tor/core/resource/ProcessRunner;Ljava/io/File;Ljava/io/File;Lkotlin/jvm/functions/Function0;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public synthetic fun <init> (Lio/matthewnelson/kmp/tor/core/resource/internal/ProcessRunner;Ljava/io/File;Ljava/io/File;Lkotlin/jvm/functions/Function0;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
 	public final fun isAndroidRuntime ()Z
 	public final fun osArch ()Lio/matthewnelson/kmp/tor/core/resource/OSArch;
 	public final fun osHost ()Lio/matthewnelson/kmp/tor/core/resource/OSHost;
@@ -106,27 +106,6 @@ public final class io/matthewnelson/kmp/tor/core/resource/PlatformResource {
 public final class io/matthewnelson/kmp/tor/core/resource/PlatformResource$Builder {
 	public field resourceClass Ljava/lang/Class;
 	public field resourcePath Ljava/lang/String;
-}
-
-public final class io/matthewnelson/kmp/tor/core/resource/ProcessExt {
-	public static final fun forciblyDestroy (Ljava/lang/Process;)Ljava/lang/Process;
-	public static final fun waitFor-8Mi8wO0 (Ljava/lang/Process;JZ)Z
-	public static synthetic fun waitFor-8Mi8wO0$default (Ljava/lang/Process;JZILjava/lang/Object;)Z
-}
-
-public abstract interface class io/matthewnelson/kmp/tor/core/resource/ProcessRunner {
-	public static final field Companion Lio/matthewnelson/kmp/tor/core/resource/ProcessRunner$Companion;
-	public static fun runAndWait (Lio/matthewnelson/kmp/tor/core/resource/ProcessRunner;Ljava/util/List;)Ljava/lang/String;
-	public abstract fun runAndWait-HG0u8IE (Ljava/util/List;J)Ljava/lang/String;
-}
-
-public final class io/matthewnelson/kmp/tor/core/resource/ProcessRunner$Companion {
-	public final fun runAndWait (Lio/matthewnelson/kmp/tor/core/resource/ProcessRunner;Ljava/util/List;)Ljava/lang/String;
-}
-
-public final class io/matthewnelson/kmp/tor/core/resource/ProcessRunner$Default : io/matthewnelson/kmp/tor/core/resource/ProcessRunner {
-	public static final field INSTANCE Lio/matthewnelson/kmp/tor/core/resource/ProcessRunner$Default;
-	public fun runAndWait-HG0u8IE (Ljava/util/List;J)Ljava/lang/String;
 }
 
 public final class io/matthewnelson/kmp/tor/core/resource/Resource {
